Added a navigation bar. Updated meal page front end.
