+++ foodplanner/lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 
 import '../pages/mealPage.dart';
+import '../pages/homePage.dart';
 
 void main() {
   runApp(const MyApp());
@@ -34,7 +35,7 @@ class MyApp extends StatelessWidget {
         useMaterial3: true,
       ),
       home:
-          const MealPage(), //const MyHomePage(title: 'Flutter Demo Home Page'),
+          const HomePage(), //const MyHomePage(title: 'Flutter Demo Home Page'),
     );
   }
 }
