+++ foodplanner/lib/pages/profilePage.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+class ProfilePage extends StatelessWidget {
+  const ProfilePage({super.key});
+  
+
+  @override Widget build(BuildContext context) {
+    return const MaterialApp();
+    
+  }
+}
