+++ foodplanner/lib/pages/settingsPage.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+class SettingsPage extends StatelessWidget {
+  const SettingsPage({super.key});
+
+  @override Widget build(BuildContext context) {
+    return const MaterialApp();
+  }
+}
